@@ -13,13 +13,13 @@ class SplitView extends StatefulWidget {
   final double initialWeight;
   final Color gripColor;
   final double positionLimit;
-  final ValueChanged<double> onWeightChanged;
+  final ValueChanged<double?>? onWeightChanged;
 
   SplitView({
-    Key key,
-    @required this.view1,
-    @required this.view2,
-    @required this.viewMode,
+    Key? key,
+    required this.view1,
+    required this.view2,
+    required this.viewMode,
     this.gripSize = 12.0,
     this.initialWeight = 0.5,
     this.gripColor = Colors.grey,
@@ -32,16 +32,17 @@ class SplitView extends StatefulWidget {
 }
 
 class _SplitViewState extends State<SplitView> {
-  double defaultWeight;
-  ValueNotifier<double> weight;
-  double _prevWeight;
+  double? defaultWeight;
+  late ValueNotifier<double?> weight;
+  double? _prevWeight;
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
 
     this.defaultWeight =
-        PageStorage.of(context)?.readState(context, identifier: widget.key) ?? widget.initialWeight;
+        PageStorage.of(context)?.readState(context, identifier: widget.key) ??
+            widget.initialWeight;
   }
 
   @override
@@ -51,18 +52,19 @@ class _SplitViewState extends State<SplitView> {
 
     return LayoutBuilder(
       builder: (context, constraints) {
-        return ValueListenableBuilder<double>(
+        return ValueListenableBuilder<double?>(
           valueListenable: weight,
           builder: (_, w, __) {
             if (widget.onWeightChanged != null && _prevWeight != w) {
               _prevWeight = w;
-              PageStorage.of(context)?.writeState(context, w, identifier: widget.key);
-              widget.onWeightChanged(w);
+              PageStorage.of(context)
+                  ?.writeState(context, w, identifier: widget.key);
+              widget.onWeightChanged!(w);
             }
             if (widget.viewMode == SplitViewMode.Vertical) {
-              return _buildVerticalView(context, constraints, w);
+              return _buildVerticalView(context, constraints, w!);
             } else {
-              return _buildHorizontalView(context, constraints, w);
+              return _buildHorizontalView(context, constraints, w!);
             }
           },
         );
@@ -70,7 +72,8 @@ class _SplitViewState extends State<SplitView> {
     );
   }
 
-  Stack _buildVerticalView(BuildContext context, BoxConstraints constraints, double w) {
+  Stack _buildVerticalView(
+      BuildContext context, BoxConstraints constraints, double w) {
     final double top = constraints.maxHeight * w;
     final double bottom = constraints.maxHeight * (1.0 - w);
     final halfGripSize = widget.gripSize / 2.0;
@@ -101,7 +104,8 @@ class _SplitViewState extends State<SplitView> {
             child: GestureDetector(
               behavior: HitTestBehavior.translucent,
               onVerticalDragUpdate: (detail) {
-                final RenderBox container = context.findRenderObject() as RenderBox;
+                final RenderBox container =
+                    context.findRenderObject() as RenderBox;
                 final pos = container.globalToLocal(detail.globalPosition);
                 if (pos.dy > widget.positionLimit &&
                     pos.dy < (container.size.height - widget.positionLimit)) {
@@ -116,7 +120,8 @@ class _SplitViewState extends State<SplitView> {
     );
   }
 
-  Widget _buildHorizontalView(BuildContext context, BoxConstraints constraints, double w) {
+  Widget _buildHorizontalView(
+      BuildContext context, BoxConstraints constraints, double w) {
     final double left = constraints.maxWidth * w;
     final double right = constraints.maxWidth * (1.0 - w);
     final double halfGripSize = widget.gripSize / 2.0;
@@ -147,7 +152,8 @@ class _SplitViewState extends State<SplitView> {
             child: GestureDetector(
               behavior: HitTestBehavior.translucent,
               onVerticalDragUpdate: (detail) {
-                final RenderBox container = context.findRenderObject() as RenderBox;
+                final RenderBox container =
+                    context.findRenderObject() as RenderBox;
                 final pos = container.globalToLocal(detail.globalPosition);
                 if (pos.dx > widget.positionLimit &&
                     pos.dx < (container.size.width - widget.positionLimit)) {
